% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslGetRateValue.R
\name{GetValue}
\alias{GetValue}
\alias{zslGetBundle}
\alias{zslGetSeries}
\alias{zslGetSingle}
\title{Get the zsl Value in Several Ways}
\usage{
zslGetSingle(code, date, ch)

zslGetSeries(code, start, end, ch, xts = TRUE, zero.fill = FALSE)

zslGetBundle(codes, start, end, ch)
}
\arguments{
\item{code}{character, the bond code in Wind sytle, which is six digits with
".SH" or ".SZ" to indicate the exchange}

\item{date}{Date, the effect date of the ratio to be obtained}

\item{ch}{the RODBC connection object}

\item{start}{the start of the period, Date object of like "YYYY-MM-DD"}

\item{end}{the end of the period, Date object of like "YYYY-MM-DD"}

\item{xts}{logical, default is TRUE, return the final series as \code{xts}
object}

\item{zero.fill}{logical, default is FALSE, if TRUE the returned series will
be filled by zero to replace those NA value while the length of the series
is equal to the period}
}
\value{
The returned value(s) are different according to the use of
  functions, see the "details."
}
\description{
Get the zsl value from the database in several ways, including get the sigle
code's single day's value, get single code's series, and so on.
}
\details{
The functions use the code to identify the bond, not the name. And the code
must be in Wind style, or with regex \code{'[0-9]{6}\\.S[HZ]'}.

For \code{zslGetSingle}, it uses params \code{code} and \code{date} to get
single value of the "code" and "date". The date should be exchange's trading
date, of course the function can also automatically find the closest trading
day before it if it is not. If there is no value returned, message will be
given, "code" and "date" should be re-checked carefully. Obviously, the
result is a single numeric value.

For \code{zslGetSeries}, it gets the series values. Surely, the params
including "start" and "end", but it can be given or missed. If both are
given, the values of specific period are returned; if only "start" or "end"
is given and the other is missed, the values after "start" or before "end"
are returned; if both are missed, all values of the bond are returned. For
the convenience, the result is an \code{xts} object as default.

If the exact period series is required, which means that the series dates
must contain all exchange trading days between "start" and "end", set param
\code{zero.fill} to \code{TRUE}. If there exists NA value, it will be
replaced by "0". This is useful when being used in \code{for} loop or
\code{*ply} functions, as the retured series all have the same length.

For \code{zslGetBundle}, it returns the "bundled" series
of more than one bond. The params \code{start} and \code{end} must be given,
and the returned the times series is an \code{xts} object of this period.
}

