% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{tidyData}
\alias{tidyData}
\title{Tidy the Data Read From File Downloaded From the Website}
\usage{
tidyData(file)
}
\arguments{
\item{file}{character, the file name must follow the rule}
}
\value{
The well-organized data.frame, ready to be written to MySQL database.
}
\description{
After downloading the file from the website, the function tidy the data read
from the file. Then organize the data to the form needed and ready to write
to the MySQL database.
}
\details{
The input file name must follow the rule, that is "date + exchange + .xls".
The file name contains two necessary element for the result, the "file date"
and the exchange. These will be two columns.

There is another important column called "serial" which will be the primary
key of database. It is organized as the following rule: "ZS" + file date +
"ST" + start date (month day) + exchange + code. Such rule will make sure
that there is no duplicated record by eliminating the duplicated rows.
}

