% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToDB.R
\name{writeToDB}
\alias{writeToDB}
\title{Write The data.frame to MySQL Database}
\usage{
writeToDB(data, channel, table, quiet = TRUE, sql.out = FALSE,
  delete.first = FALSE, primary = 1L)
}
\arguments{
\item{data}{the data.frame will be written to the database}

\item{channel}{an ODBC channel object}

\item{table}{the table name of database, which is used to organized query}

\item{quiet}{logical, default is TRUE, show the progress bar}

\item{sql.out}{logical, whether give the SQL message, just in case there are
errors occur}

\item{delete.first}{logical, whether to delete the records first using the
primary key column}

\item{primary}{integer, which column is the primary column, and just used
when delete records}
}
\value{
Besides errors messages, no return.
}
\description{
Write the data.frame to MySQL database using the method that sent the query
to the database server.
}
\details{
The function can delete the records before writing to the database, however
if it can be make sure that the there is no duplicated records, the deletion
can be skipped, saving time. The deletion work is based on the primary key,
so the column index has to be given, and the column name should the same as
that in the database.

The mechanism of the function is to **paste** an INSERT query and send to
the database server. Sum method seems safe to be executed
}

