% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslAddBondInfo.R
\name{zslAddBondInfo}
\alias{zslAddBondInfo}
\title{Adding Bond Information to Given data.frame}
\usage{
zslAddBondInfo(data)
}
\arguments{
\item{data}{the given \code{data.frame}, including "code" column}
}
\value{
A \code{data.frame} which is \code{cbind}ed with original "data" and
  bond information columns.
}
\description{
The function adds the basic bond information to the given data.frame, using
data's "code" column. So, a "code" column of Wind style bond code vector is
necessary in the given data.
}
\details{
The bond information added includes bond's maturity, bond rating, issuer
rating, rating change, bond type, industry, municipal or not. The function
will process the data retrieved from \code{WindR} Package, especially, most
information has no value for the risk-free bonds, e.g. treasury bond, local
government bond, and so on.
}

