% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslCheckMissingDate.R
\name{zslCheckMissingDate}
\alias{zslCheckMissingDate}
\title{Checking The Missing Date}
\usage{
zslCheckMissingDate(data, code, start, end)
}
\arguments{
\item{data}{the data.frame is going to be checked, must have a "date" column}

\item{code}{the code, only length 1 character vector}

\item{start}{the "start" date of the exchange date series}

\item{end}{the "end" date of the exchange date series}
}
\value{
If there is no missing date, return FALSE, otherwise, return the
  missing date(s) series.
}
\description{
Checking the missing date of given data.frame, usually from database to check
the completeness of the zsl ratio. The data.frame given \bold{MUST} have a
"date" column.
}
\details{
The param \code{code} can be given or missed, it depends on whether the given
data is filtered, so it is just a filtering element. Then "start" and "end"
are like it, if given the exchange date series is influenced.

The mechanism is that, the function find the "range" of the "date" column,
and use this range to get the exchange trading date series. After comparing
two "Date" series, the difference is returned.
}

