% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslCombineList.R
\name{zslCombineList}
\alias{zslCombineList}
\title{Get the File List From Now to the Given Date}
\usage{
zslCombineList(start)
}
\arguments{
\item{start}{the start date by which the function will get all files after
this date}
}
\value{
The data.frame contains all files from now to the given date.
}
\description{
This function get the full file list from the website from now to the given
date.
}
\details{
The given date should not be earlier than May 18th, 2005, that is the start
date of the announcement of the discount rate. If the start date isn't given,
the first page of file list is returned.
}

