% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslDownList.R
\name{zslDownList}
\alias{zslDownList}
\title{Download the URL List of the zsl Data}
\usage{
zslDownList(url)
}
\arguments{
\item{url}{the url to be parsed and get the information}
}
\value{
A data.frame contains the announcement date, exchange,
  file name and url of the file.
}
\description{
Check the website and scape the urls and other attibutes, then form the
result data.frame containing the infomation for the further use.
}
\details{
This function mostly use the \code{xml2} package to get information from the
website. The final data.frame contains the announcement date, exchange,
file name and url of the file. There is only one param \code{url}, which is
used to be checked and get the corresponding info. The zsl list is in several
pages (now nearly 150 pages), the default \code{url} is the first page where
the newest files' urls can be get.

The file name contained in the result data.frame is strictly organized. That
is "date + exchange + .xls". For the further use, this is the basic rule.
}

