% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslPeriodDiff.R
\name{zslPeriodDiff}
\alias{zslPeriodDiff}
\title{Compute the zsl Change of a Period}
\usage{
zslPeriodDiff(day1, day2, ch, scale = 0)
}
\arguments{
\item{day1}{the start of the period, Date object of like "YYYY-MM-DD"}

\item{day2}{the end of the period, Date object of like "YYYY-MM-DD"}

\item{ch}{RODBC connection object}

\item{scale}{the scale of the change, default is 0, that's meaning it does
not work, see the details of \link{\code{zslTwoDayDiff}}}
}
\value{
Tibble data.frame containing the change of the period.
}
\description{
The input are the days of the beginning and end of the period, then obtain
the change of zsl of this period, that is the difference between the first
day and last day's data. Also, the result provides the days of the period for
each code.
}
\details{
First of all, the records with "0" ratio are deleted, because those are
going to be unlisted from the market, not worth consideration. Then, not all
the bonds have more than two days' records, therefore, those with one record
are thought to be unqualified for period computation, and deleted.

For the result, only the bonds with change other than "0" will be reserved.
}

