% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslDownFile.R
\name{zslDownFile}
\alias{zslDownFile}
\title{Download the zsl Files}
\usage{
zslDownFile(dl, folder, overwrite = FALSE)
}
\arguments{
\item{dl}{the download file list, containing the file name and urls}

\item{folder}{the folder saving the files, can ben missed}

\item{overwrite}{logical, if TRUE any existing file will be overwritten}
}
\value{
No return, the file will be downloaded.
}
\description{
Download the zsl files by the file list given by \code{zslDownList} or
\code{zslCombineList}. A folder name can be given to save the files, if which
does not exist it will be created.
}

