% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zslUpdateDB.R
\name{zslUpdateDB}
\alias{zslUpdateDB}
\title{Update zsl Data}
\usage{
zslUpdateDB(ch, date, folder = "", delete.file = TRUE)
}
\arguments{
\item{delete.file}{}
}
\description{
Update the bond discount rate data to MySQL database. The date of the lastest
records can be detected from the database, and search the new data from the
website, so the param \code{date} can be ignored.
}

